# generated/node/src/routes/alertChannels.routes.ts
// @generated by flusk-lang — DO NOT EDIT

import { type FastifyInstance } from 'fastify';
import type Database from 'better-sqlite3';
import { listAlertChannels } from '../functions/list-alert-channels.function.js';
import { createAlertChannel } from '../functions/create-alert-channel.function.js';
import { findAlertChannelById } from '../functions/find-alert-channel-by-id.function.js';
import { updateAlertChannel } from '../functions/update-alert-channel.function.js';
import { deleteAlertChannel } from '../functions/delete-alert-channel.function.js';

interface AppWithDb extends FastifyInstance { db: Database.Database; }

export const alertChannelsRoutes = async (app: FastifyInstance): Promise<void> => {
  const { db } = app as AppWithDb;

  app.get('/api/alert-channels', async (request, reply) => {
    try {
      const result = listAlertChannels(db);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'listAlertChannels failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.post('/api/alert-channels', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = createAlertChannel(db, body);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'createAlertChannel failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/alert-channels/:id', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = findAlertChannelById(db, id);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'findAlertChannelById failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.put('/api/alert-channels/:id', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = updateAlertChannel(db, body);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'updateAlertChannel failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.delete('/api/alert-channels/:id', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = deleteAlertChannel(db, id);
      return reply.status(204).send();
    } catch (err: unknown) {
      request.log.error(err, 'deleteAlertChannel failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

};