# generated/node/src/routes/apiKeys.routes.ts
// @generated by flusk-lang — DO NOT EDIT

import { type FastifyInstance } from 'fastify';
import type Database from 'better-sqlite3';
import { createApiKey } from '../functions/create-api-key.function.js';
import { listApiKeys } from '../functions/list-api-keys.function.js';
import { revokeApiKey } from '../functions/revoke-api-key.function.js';

interface AppWithDb extends FastifyInstance { db: Database.Database; }

export const apiKeysRoutes = async (app: FastifyInstance): Promise<void> => {
  const { db } = app as AppWithDb;

  app.post('/api/api-keys', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = createApiKey(db, body);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'createApiKey failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/api-keys', async (request, reply) => {
    try {
      const result = listApiKeys(db);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'listApiKeys failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.post('/api/api-keys/:id/revoke', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = revokeApiKey(db, body);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'revokeApiKey failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

};