# generated/node/src/routes/budgetAlerts.routes.ts
// @generated by flusk-lang — DO NOT EDIT

import { type FastifyInstance } from 'fastify';
import type Database from 'better-sqlite3';
import { listBudgetAlerts } from '../functions/list-budget-alerts.function.js';
import { createBudgetAlert } from '../functions/create-budget-alert.function.js';
import { findBudgetAlertById } from '../functions/find-budget-alert-by-id.function.js';
import { updateBudgetAlert } from '../functions/update-budget-alert.function.js';
import { deleteBudgetAlert } from '../functions/delete-budget-alert.function.js';
import { getBudgetStatus } from '../functions/get-budget-status.function.js';
import { checkBudget } from '../functions/check-budget.function.js';

interface AppWithDb extends FastifyInstance { db: Database.Database; }

export const budgetAlertsRoutes = async (app: FastifyInstance): Promise<void> => {
  const { db } = app as AppWithDb;

  app.get('/api/budget-alerts', async (request, reply) => {
    try {
      const result = listBudgetAlerts(db);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'listBudgetAlerts failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.post('/api/budget-alerts', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = createBudgetAlert(db, body);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'createBudgetAlert failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/budget-alerts/:id', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = findBudgetAlertById(db, id);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'findBudgetAlertById failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.put('/api/budget-alerts/:id', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = updateBudgetAlert(db, body);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'updateBudgetAlert failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.delete('/api/budget-alerts/:id', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = deleteBudgetAlert(db, id);
      return reply.status(204).send();
    } catch (err: unknown) {
      request.log.error(err, 'deleteBudgetAlert failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/budget-alerts/status', async (request, reply) => {
    try {
      const result = getBudgetStatus(db);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'getBudgetStatus failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.post('/api/budget-alerts/check', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = checkBudget(db, body);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'checkBudget failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

};