# generated/node/src/routes/costAttributions.routes.ts
// @generated by flusk-lang — DO NOT EDIT

import { type FastifyInstance } from 'fastify';
import type Database from 'better-sqlite3';
import { listCostAttributions } from '../functions/list-cost-attributions.function.js';
import { createCostAttribution } from '../functions/create-cost-attribution.function.js';
import { getCostByDimension } from '../functions/get-cost-by-dimension.function.js';
import { getCostTrend } from '../functions/get-cost-trend.function.js';
import { getTopSpenders } from '../functions/get-top-spenders.function.js';

interface AppWithDb extends FastifyInstance { db: Database.Database; }

export const costAttributionsRoutes = async (app: FastifyInstance): Promise<void> => {
  const { db } = app as AppWithDb;

  app.get('/api/cost-attributions', async (request, reply) => {
    try {
      const result = listCostAttributions(db);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'listCostAttributions failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.post('/api/cost-attributions', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = createCostAttribution(db, body);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'createCostAttribution failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/cost-attributions/by-dimension/:dimension', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = getCostByDimension(db, id);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'getCostByDimension failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/cost-attributions/trend/:dimension/:value', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = getCostTrend(db, id);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'getCostTrend failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/cost-attributions/top-spenders/:dimension', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = getTopSpenders(db, id);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'getTopSpenders failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

};