# generated/node/src/routes/dashboardWidgets.routes.ts
// @generated by flusk-lang — DO NOT EDIT

import { type FastifyInstance } from 'fastify';
import type Database from 'better-sqlite3';
import { listDashboardWidgets } from '../functions/list-dashboard-widgets.function.js';
import { createDashboardWidget } from '../functions/create-dashboard-widget.function.js';
import { findDashboardWidgetById } from '../functions/find-dashboard-widget-by-id.function.js';
import { updateDashboardWidget } from '../functions/update-dashboard-widget.function.js';
import { deleteDashboardWidget } from '../functions/delete-dashboard-widget.function.js';

interface AppWithDb extends FastifyInstance { db: Database.Database; }

export const dashboardWidgetsRoutes = async (app: FastifyInstance): Promise<void> => {
  const { db } = app as AppWithDb;

  app.get('/api/dashboard-widgets', async (request, reply) => {
    try {
      const result = listDashboardWidgets(db);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'listDashboardWidgets failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.post('/api/dashboard-widgets', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = createDashboardWidget(db, body);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'createDashboardWidget failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/dashboard-widgets/:id', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = findDashboardWidgetById(db, id);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'findDashboardWidgetById failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.put('/api/dashboard-widgets/:id', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = updateDashboardWidget(db, body);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'updateDashboardWidget failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.delete('/api/dashboard-widgets/:id', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = deleteDashboardWidget(db, id);
      return reply.status(204).send();
    } catch (err: unknown) {
      request.log.error(err, 'deleteDashboardWidget failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

};