# generated/openapi.yaml
openapi: 3.1.0
info:
  title: Flusk API
  version: 0.1.0
paths:
...
  /api/alert-channels/:
    get:
      operationId: listAlertChannels
      tags:
        - alertChannels
      summary: List all alert channels
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertChannel'
    post:
      operationId: createAlertChannel
      tags:
        - alertChannels
      summary: Create a new alert channel
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertChannel'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertChannel'
  /api/alert-channels/:id:
    get:
      operationId: findAlertChannelById
      tags:
        - alertChannels
      summary: Get channel by ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertChannel'
    put:
      operationId: updateAlertChannel
      tags:
        - alertChannels
      summary: Update a channel
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertChannel'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertChannel'
    delete:
      operationId: deleteAlertChannel
      tags:
        - alertChannels
      summary: Delete a channel
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertChannel'
...
  /api/api-keys/:
    post:
      operationId: createApiKey
      tags:
        - apiKeys
      summary: Create API key (returns plaintext key once)
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKey'
    get:
      operationId: listApiKeys
      tags:
        - apiKeys
      summary: List API keys (shows prefix only)
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
  /api/api-keys/:id/revoke:
    post:
      operationId: revokeApiKey
      tags:
        - apiKeys
      summary: Revoke an API key
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKey'
...
  /api/budget-alerts/:
    get:
      operationId: listBudgetAlerts
      tags:
        - budgetAlerts
      summary: List budget alerts
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetAlert'
    post:
      operationId: createBudgetAlert
      tags:
        - budgetAlerts
      summary: Create a budget alert
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetAlert'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BudgetAlert'
  /api/budget-alerts/:id:
    get:
      operationId: findBudgetAlertById
      tags:
        - budgetAlerts
      summary: Get budget alert by ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetAlert'
    put:
      operationId: updateBudgetAlert
      tags:
        - budgetAlerts
      summary: Update a budget alert
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetAlert'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BudgetAlert'
    delete:
      operationId: deleteBudgetAlert
      tags:
        - budgetAlerts
      summary: Delete a budget alert
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetAlert'
...
  /api/budget-alerts/check:
    post:
      operationId: checkBudget
      tags:
        - budgetAlerts
      summary: Run budget check now
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetAlert'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BudgetAlert'
...
  /api/cost-attributions/:
    get:
      operationId: listCostAttributions
      tags:
        - costAttributions
      summary: List all cost attributions
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostAttribution'
    post:
      operationId: createCostAttribution
      tags:
        - costAttributions
      summary: Create attribution
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostAttribution'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CostAttribution'
...
  /api/dashboard-widgets/:
    get:
      operationId: listDashboardWidgets
      tags:
        - dashboardWidgets
      summary: List widgets (filter by dashboardId)
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardWidget'
    post:
      operationId: createDashboardWidget
      tags:
        - dashboardWidgets
      summary: Create a new widget
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardWidget'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DashboardWidget'
  /api/dashboard-widgets/:id:
    get:
      operationId: findDashboardWidgetById
      tags:
        - dashboardWidgets
      summary: Get widget by ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardWidget'
    put:
      operationId: updateDashboardWidget
      tags:
        - dashboardWidgets
      summary: Update a widget
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardWidget'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DashboardWidget'
    delete:
      operationId: deleteDashboardWidget
      tags:
        - dashboardWidgets
      summary: Delete a widget
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardWidget'
...
components:
  schemas:
...
    AlertChannel:
      type: object
      properties:
        name:
          type: string
          description: Human-readable channel name
        channelType:
          $ref: '#/components/schemas/enum'
          description: Alert provider type
          enum:
            - pagerduty
            - slack
            - discord
            - webhook
            - email
        config:
          type: object
          description: Provider-specific configuration (routing_key, webhook_url, etc.)
        enabled:
          type: boolean
          description: Whether this channel is active
        severityFilter:
          $ref: '#/components/schemas/enum'
          description: Minimum severity level to trigger alerts on this channel
          enum:
            - critical
            - warning
            - info
      required:
        - name
        - channelType
        - config
        - enabled
        - severityFilter
      description: Configuration for an alert destination (PagerDuty, Slack, etc.)
...
    ApiKey:
      type: object
      properties:
        key:
          type: string
          description: SHA-256 hashed key (never store plaintext)
        prefix:
          type: string
          description: First 8 chars of key for display
        name:
          type: string
          description: Human-readable key name
        organizationId:
          type: string
          description: Owning organization
        userId:
          type: string
          description: User who created the key
        scopes:
          type: object
          description: Allowed scopes, e.g. ["ingest","read","admin"]
        expiresAt:
          type: string
          description: Key expiration timestamp
        lastUsedAt:
          type: string
          description: Last time the key was used
        status:
          $ref: '#/components/schemas/enum'
          description: Key status
          enum:
            - active
            - revoked
      required:
        - key
        - prefix
        - name
        - organizationId
        - userId
        - scopes
        - expiresAt
        - lastUsedAt
        - status
      description: Hashed API key for programmatic access
...
    BudgetAlert:
      type: object
      properties:
        name:
          type: string
          description: Alert name
        budgetType:
          $ref: '#/components/schemas/enum'
          description: Budget period type
          enum:
            - daily
            - weekly
            - monthly
        thresholdUsd:
          type: number
          description: Spending threshold in USD
        warningPercent:
          type: integer
          description: Percentage of threshold to trigger warning
        currentSpendUsd:
          type: number
          description: Current spend in this period
        enabled:
          type: boolean
          description: Whether this alert is active
        autoPause:
          type: boolean
          description: Auto-pause agents when exceeded
        lastTriggeredAt:
          type: string
          description: ISO datetime of last trigger
        agentFilter:
          type: string
          description: Optional agent name filter (null = all agents)
      required:
        - name
        - budgetType
        - thresholdUsd
        - warningPercent
        - currentSpendUsd
        - enabled
        - autoPause
        - lastTriggeredAt
        - agentFilter
      description: Budget threshold alert configuration
...
    CostAttribution:
      type: object
      properties:
        llmCallId:
          type: string
          description: Reference to the LLM call
        tagKey:
          type: string
          description: Dimension (feature, team, customer)
        tagValue:
          type: string
          description: Value (support-bot, engineering, acme-corp)
        costUsd:
          type: number
          description: Cost attributed to this tag
        inputTokens:
          type: integer
          description: Input tokens for this call
        outputTokens:
          type: integer
          description: Output tokens for this call
        model:
          type: string
          description: Model used
        provider:
          type: string
          description: Provider (openai, anthropic)
      required:
        - llmCallId
        - tagKey
        - tagValue
        - costUsd
        - inputTokens
        - outputTokens
        - model
        - provider
      description: Links an LLM call to cost tags for attribution tracking
...
    DashboardWidget:
      type: object
      properties:
        dashboardId:
          type: string
          description: Parent dashboard ID
        widgetType:
          $ref: '#/components/schemas/enum'
          description: Type of visualization widget
          enum:
            - line-chart
            - bar-chart
            - pie-chart
            - metric-card
            - table
            - heatmap
            - waterfall
            - timeline
        title:
          type: string
          description: Widget display title
        query:
          type: object
          description: Data query config (entity, aggregation, groupBy, timeRange)
        position:
          type: object
          description: Grid position {x, y, w, h}
        config:
          type: object
          description: Widget-specific config (colors, thresholds, labels)
        refreshOverride:
          type: integer
          description: Override dashboard refresh interval for this widget
      required:
        - dashboardId
        - widgetType
        - title
        - query
        - position
        - config
        - refreshOverride
      description: Individual widget within a dashboard
...